Initialize objects in dnxRegistrar before using.
git-svn-id: https://svn.forschooner.net/code/dnx/trunk/dnx@707 2be5450d-c085-40c6-83eb-9dc9c13f2dba

+++ server/dnxTimer.c
@@ -122,7 +122,7 @@ static void * dnxTimer(void * data)
       // harmonic where the dnxClient and timer thread are both expiring at the
       // same time.
       srand(time(0));
-      skew = (rand() % itimer->sleepms) - (itimer->sleepms / 2);
+      skew = (rand() % itimer->sleepms) - (int)(itimer->sleepms / 2);
    
       dnxDebug(2, "dnxTimerCreate: Skew is (%d)", itimer->sleepms + skew);
